application and fx controller separation
